SPELLMEM...DESC will now return the INFO.DESC value if it exists.  The GUI for all objects will return the INFO.DESC value if it exists, otherwise the normal DESC value.
