+++ code/src/java/pcgen/core/PlayerCharacter.java
@@ -17,6 +17,7 @@ package pcgen.core;
 
 import java.awt.Rectangle;
 import java.math.BigDecimal;
+import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -8179,6 +8180,9 @@ public class PlayerCharacter implements Cloneable, VariableContainer
 		if (b instanceof PObject)
 		{
 			cdo = (PObject) b;
+			String dString = getInfoToken(".INFO.DESC", cdo);
+			if (!dString.equals(".INFO.DESC"))
+				return dString;
 		}
 		else if (b instanceof CNAbility)
 		{
@@ -8212,6 +8216,29 @@ public class PlayerCharacter implements Cloneable, VariableContainer
 		}
 		return sb.toString();
 	}
+
+	public String getInfoToken(String token, PObject po) {
+		// looking for a token in the form of RACE.INFO.TAG where
+		// RACE indicate which token map to check for a INFO label of TAG to return
+		int i = token.indexOf(".INFO.");
+		String ts = token;
+		if (i>-1)
+			ts = token.substring(i+6).toUpperCase();
+		else
+			return token;
+		Set<MapKey<?, ?>> keys = po.getMapKeys();
+		for (MapKey<?, ?> key : keys) {
+			Map<?, ?> key2 = po.getMapFor(key);
+			for(Object k : key2.keySet()) {
+				if (k.toString().equals(ts)) {
+					MessageFormat m = (MessageFormat) key2.get(k);
+					return m.toPattern();
+				}
+			}
+		}
+		return token;
+	}
+
 
 	/**
 	 * This method gets the information about the levels at which classes and
